coastcheck: Improve accuracy of processing, previously the data was rounded to +-2m by the float precision, using a double fixes this, at the expense of a little more memory.

+++ applications/utils/coastcheck/osmtypes.h
@@ -6,8 +6,8 @@
 
 struct osmNode {
     int id;
-    float lon;
-    float lat;
+    double lon;
+    double lat;
 };
 
 struct osmWay {
